requirements/base.txt
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-Script==2.0.6
GeoAlchemy2==0.6.3
marshmallow==3.5.0
psycopg2-binary==2.8.4
Shapely==1.7.0
gunicorn==20.0.4
python-decouple==3.3